terms js -> tsx
